+++ Changelog.md
@@ -12,7 +12,7 @@ Version 1.7 Beta (Sat 06 Jul 2013)
 * Renaming Tracks using a user specified list of names
 * Allow searching for a album using an URL
 * Support rarer languages such as English (Official)  and English (Localized)
-* Length field formatted as time
+* Length field formatted as time.
 * ISRC field
 
 ### Improvements ###